Split InternalObjectsSubscriber into 2. One per object
