+++ EventListener/CompanyObjectSubscriber.php
@@ -21,19 +21,12 @@ use MauticPlugin\IntegrationsBundle\Event\InternalObjectRouteEvent;
 use MauticPlugin\IntegrationsBundle\Event\InternalObjectUpdateEvent;
 use MauticPlugin\IntegrationsBundle\IntegrationEvents;
 use MauticPlugin\IntegrationsBundle\Internal\Object\Company;
-use MauticPlugin\IntegrationsBundle\Internal\Object\Contact;
 use MauticPlugin\IntegrationsBundle\Sync\SyncDataExchange\Internal\ObjectHelper\CompanyObjectHelper;
-use MauticPlugin\IntegrationsBundle\Sync\SyncDataExchange\Internal\ObjectHelper\ContactObjectHelper;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Routing\Router;
 
-class InternalObjectSubscriber implements EventSubscriberInterface
+class CompanyObjectSubscriber implements EventSubscriberInterface
 {
-    /**
-     * @var ContactObjectHelper
-     */
-    private $contactObjectHelper;
-
     /**
      * @var CompanyObjectHelper
      */
@@ -45,16 +38,13 @@ class InternalObjectSubscriber implements EventSubscriberInterface
     private $router;
 
     /**
-     * @param ContactObjectHelper $contactObjectHelper
      * @param CompanyObjectHelper $companyObjectHelper
      * @param Router              $router
      */
     public function __construct(
-        ContactObjectHelper $contactObjectHelper,
         CompanyObjectHelper $companyObjectHelper,
         Router $router
     ) {
-        $this->contactObjectHelper = $contactObjectHelper;
         $this->companyObjectHelper = $companyObjectHelper;
         $this->router              = $router;
     }
@@ -66,30 +56,15 @@ class InternalObjectSubscriber implements EventSubscriberInterface
     {
         return [
             IntegrationEvents::INTEGRATION_COLLECT_INTERNAL_OBJECTS => ['collectInternalObjects', 0],
-            IntegrationEvents::INTEGRATION_UPDATE_INTERNAL_OBJECTS  => [
-                ['updateContacts', 0],
-                ['updateCompanies', 0],
-            ],
-            IntegrationEvents::INTEGRATION_CREATE_INTERNAL_OBJECTS => [
-                ['createContacts', 0],
-                ['createCompanies', 0],
-            ],
-            IntegrationEvents::INTEGRATION_FIND_INTERNAL_RECORDS => [
-                ['findContactsByIds', 0],
-                ['findContactsByDateRange', 0],
-                ['findContactsByFieldValues', 0],
+            IntegrationEvents::INTEGRATION_UPDATE_INTERNAL_OBJECTS  => ['updateCompanies', 0],
+            IntegrationEvents::INTEGRATION_CREATE_INTERNAL_OBJECTS  => ['createCompanies', 0],
+            IntegrationEvents::INTEGRATION_FIND_INTERNAL_RECORDS    => [
                 ['findCompaniesByIds', 0],
                 ['findCompaniesByDateRange', 0],
                 ['findCompaniesByFieldValues', 0],
             ],
-            IntegrationEvents::INTEGRATION_FIND_OWNER_IDS => [
-                ['findOwnerIdsForContacts', 0],
-                ['findOwnerIdsForCompanies', 0],
-            ],
-            IntegrationEvents::INTEGRATION_BUILD_INTERNAL_OBJECT_ROUTE => [
-                ['buildContactRoute', 0],
-                ['buildCompanyRoute', 0],
-            ],
+            IntegrationEvents::INTEGRATION_FIND_OWNER_IDS              => ['findOwnerIdsForCompanies', 0],
+            IntegrationEvents::INTEGRATION_BUILD_INTERNAL_OBJECT_ROUTE => ['buildCompanyRoute', 0],
         ];
     }
 
@@ -98,26 +73,7 @@ class InternalObjectSubscriber implements EventSubscriberInterface
      */
     public function collectInternalObjects(InternalObjectEvent $event): void
     {
-        $event->addObject(new Contact());
         $event->addObject(new Company());
-    }
-
-    /**
-     * @param InternalObjectUpdateEvent $event
-     */
-    public function updateContacts(InternalObjectUpdateEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName()) {
-            return;
-        }
-
-        $event->setUpdatedObjectMappings(
-            $this->contactObjectHelper->update(
-                $event->getIdentifiedObjectIds(),
-                $event->getUpdateObjects()
-            )
-        );
-        $event->stopPropagation();
     }
 
     /**
@@ -130,7 +86,7 @@ class InternalObjectSubscriber implements EventSubscriberInterface
         }
 
         $event->setUpdatedObjectMappings(
-            $this->contactObjectHelper->update(
+            $this->companyObjectHelper->update(
                 $event->getIdentifiedObjectIds(),
                 $event->getUpdateObjects()
             )
@@ -141,39 +97,13 @@ class InternalObjectSubscriber implements EventSubscriberInterface
     /**
      * @param InternalObjectCreateEvent $event
      */
-    public function createContacts(InternalObjectCreateEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName()) {
-            return;
-        }
-
-        $event->setObjectMappings($this->contactObjectHelper->create($event->getCreateObjects()));
-        $event->stopPropagation();
-    }
-
-    /**
-     * @param InternalObjectCreateEvent $event
-     */
     public function createCompanies(InternalObjectCreateEvent $event): void
     {
-        if (Contact::NAME !== $event->getObject()->getName()) {
+        if (Company::NAME !== $event->getObject()->getName()) {
             return;
         }
 
         $event->setObjectMappings($this->companyObjectHelper->create($event->getCreateObjects()));
-        $event->stopPropagation();
-    }
-
-    /**
-     * @param InternalObjectFindEvent $event
-     */
-    public function findContactsByIds(InternalObjectFindEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName() || empty($event->getIds())) {
-            return;
-        }
-
-        $event->setFoundObjects($this->contactObjectHelper->findObjectsByIds($event->getIds()));
         $event->stopPropagation();
     }
 
@@ -187,26 +117,6 @@ class InternalObjectSubscriber implements EventSubscriberInterface
         }
 
         $event->setFoundObjects($this->companyObjectHelper->findObjectsByIds($event->getIds()));
-        $event->stopPropagation();
-    }
-
-    /**
-     * @param InternalObjectFindEvent $event
-     */
-    public function findContactsByDateRange(InternalObjectFindEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName() || empty($event->getDateRange())) {
-            return;
-        }
-
-        $event->setFoundObjects(
-            $this->contactObjectHelper->findObjectsBetweenDates(
-                $event->getDateRange()->getFromDate(),
-                $event->getDateRange()->getToDate(),
-                $event->getStart(),
-                $event->getLimit()
-            )
-        );
         $event->stopPropagation();
     }
 
@@ -233,23 +143,6 @@ class InternalObjectSubscriber implements EventSubscriberInterface
     /**
      * @param InternalObjectFindEvent $event
      */
-    public function findContactsByFieldValues(InternalObjectFindEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName() || empty($event->getFieldValues())) {
-            return;
-        }
-
-        $event->setFoundObjects(
-            $this->contactObjectHelper->findObjectsByFieldValues(
-                $event->getFieldValues()
-            )
-        );
-        $event->stopPropagation();
-    }
-
-    /**
-     * @param InternalObjectFindEvent $event
-     */
     public function findCompaniesByFieldValues(InternalObjectFindEvent $event): void
     {
         if (Company::NAME !== $event->getObject()->getName() || empty($event->getFieldValues())) {
@@ -267,23 +160,6 @@ class InternalObjectSubscriber implements EventSubscriberInterface
     /**
      * @param InternalObjectOwnerEvent $event
      */
-    public function findOwnerIdsForContacts(InternalObjectOwnerEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName()) {
-            return;
-        }
-
-        $event->setOwners(
-            $this->contactObjectHelper->findOwnerIds(
-                $event->getObjectIds()
-            )
-        );
-        $event->stopPropagation();
-    }
-
-    /**
-     * @param InternalObjectOwnerEvent $event
-     */
     public function findOwnerIdsForCompanies(InternalObjectOwnerEvent $event): void
     {
         if (Company::NAME !== $event->getObject()->getName()) {
@@ -293,27 +169,6 @@ class InternalObjectSubscriber implements EventSubscriberInterface
         $event->setOwners(
             $this->companyObjectHelper->findOwnerIds(
                 $event->getObjectIds()
-            )
-        );
-        $event->stopPropagation();
-    }
-
-    /**
-     * @param InternalObjectRouteEvent $event
-     */
-    public function buildContactRoute(InternalObjectRouteEvent $event): void
-    {
-        if (Contact::NAME !== $event->getObject()->getName()) {
-            return;
-        }
-
-        $event->setRoute(
-            $this->router->generate(
-                'mautic_contact_action',
-                [
-                    'objectAction' => 'view',
-                    'objectId'     => $event->getId(),
-                ]
             )
         );
         $event->stopPropagation();
